requirements.txt
moviepy>=1.0.3
pydub>=0.25.1
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
gspread>=5.10.0
